Move login details validation to provider
