+++ pkg/provider/akamai/akamai.go
@@ -49,6 +49,8 @@ type MfaUserOption struct {
 
 // Client is a wrapper representing a Akamai SAML client
 type Client struct {
+	provider.ValidateBase
+
 	client *provider.HTTPClient
 	mfa    string
 }
