+++ pkg/provider/base_validate_test.go
@@ -1,4 +1,4 @@
-package saml2aws
+package provider
 
 import (
 	"testing"
@@ -30,7 +30,8 @@ func TestLoginDetails_Validate(t *testing.T) {
 				Password: tt.fields.Password,
 				URL:      tt.fields.URL,
 			}
-			if err := ld.Validate(); (err != nil) != tt.wantErr {
+			providerstub := &ValidateBase{}
+			if err := providerstub.Validate(ld); (err != nil) != tt.wantErr {
 				t.Errorf("LoginDetails.Validate() error = %v, wantErr %v", err, tt.wantErr)
 			}
 		})
