+++ pkg/provider/f5apm/f5apm.go
@@ -26,6 +26,8 @@ var logger = logrus.WithField("provider", "f5apm")
 
 //Client client for F5 APM
 type Client struct {
+	provider.ValidateBase
+
 	client   *provider.HTTPClient
 	policyID string
 }
